+++ miamoremustachio/ToDo/modules/services/user-services.js
@@ -1,4 +1,4 @@
-const { User } = require('../models/user.js');
+const { User } = require('../models/user-model.js');
 
 const users = {
   model: User,
